commited update

+++ ft_numhexlen.c
@@ -6,13 +6,13 @@
 /*   By: carbon-m <carbon-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 15:25:24 by carbon-m          #+#    #+#             */
-/*   Updated: 2024/11/04 16:56:42 by carbon-m         ###   ########.fr       */
+/*   Updated: 2024/11/05 16:16:08 by carbon-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_numhexlen(unsigned int input)
+int	ft_numhexlen(unsigned long long input)
 {
 	int	i;
