+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: carbon-m <carbon-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/01 17:27:06 by carbon-m          #+#    #+#             */
-/*   Updated: 2024/11/04 15:26:33 by carbon-m         ###   ########.fr       */
+/*   Updated: 2024/11/05 16:19:32 by carbon-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int		ft_putchar(int input);
 int		ft_putstr(char *input);
 int		ft_putnbr(int input);
 int		ft_putunsnbr(unsigned int input);
-int		ft_numhexlen(unsigned int input);
+int		ft_numhexlen(unsigned long long input);
 int		ft_puthex(unsigned int input);
 int		ft_puthexupper(unsigned int input);
 int		ft_putmem(unsigned long long input);
